refactored navigator
