+++ src/screens/index.ts
@@ -1,13 +1,11 @@
 import AllExpensesScreen from "./AllExpenses";
 import MonthlyExpensesScreen from "./MonthlyExpenses";
 import ExpenseFormScreen from "./ExpenseForm";
-import Navigator from "./Navigator";
 import LocaleSelectorScreen from "./LocaleSelector";
 
 export {
   AllExpensesScreen,
   MonthlyExpensesScreen,
   ExpenseFormScreen,
-  Navigator,
   LocaleSelectorScreen,
 };
